requirements.txt
pandas==2.1.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
